Reorganize and do simple plugin test
