+++ build_and_run_sst.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+docker build . -t sst-plugin -f plugins/sst/Dockerfile
+docker run --rm -it sst-plugin:latest
