+++ build_and_run_temp_precip.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+docker build . -t temp-precip-plugin -f plugins/temp_precip/Dockerfile
+docker run --rm -it temp-precip-plugin:latest
