+++ extract_storms_v2.py
@@ -9,7 +9,6 @@ import enum
 import json
 import logging
 import os
-
 from datetime import datetime
 from typing import Any, Tuple
