+++ ms/create_hms_grid.py
@@ -3,15 +3,15 @@ import os
 from datetime import datetime
 from typing import Union
 
+import boto3
 import pyproj
 from constants import INDEX
+from meilisearch import Client
 from pydsstools.heclib.dss import HecDss
 from pydsstools.heclib.utils import SHG_WKT
 from shapely.geometry import Point
 from shapely.ops import transform
 from storm_query import query_ms
-
-from ms.client_utils import create_meilisearch_client, create_s3_client
 
 
 def strip_all_whitespace(dirty_string: str) -> str:
@@ -62,8 +62,9 @@ def main(
     limit: int = 1000,
 ):
     # Create s3 and meilisearch clients from environment variables
-    s3_client = create_s3_client(access_key_id, secret_access_key)
-    ms_client = create_meilisearch_client(ms_host, ms_api_key)
+    session = boto3.session.Session(access_key_id, secret_access_key)
+    s3_client = session.client("s3")
+    ms_client = Client(ms_host, ms_api_key)
 
     # Create data directory from watershed if none provided
     if data_directory == None:
