+++ plugins/sst/entrypoint.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+
+echo 'Stochaistic storm transposition modeling plugin'
+python -m main
