+++ plugins/sst/main.py
@@ -6,7 +6,7 @@ import sys
 
 from papipyplug import parse_input, plugin_logger, print_results
 
-from .pull_temp_precip_plugin_module import PLUGIN_PARAMS, main
+from sst_plugin import PLUGIN_PARAMS, main
 
 if __name__ == "__main__":
     # Start plugin logger
