+++ plugins/sst/requirements.txt
@@ -1,0 +1,51 @@
+affine==2.4.0
+attrs==23.1.0
+boto3==1.22.2
+botocore==1.25.13
+certifi==2023.7.22
+click==8.1.7
+click-plugins==1.1.1
+cligj==0.7.2
+cycler==0.12.1
+dataclasses-json==0.5.7
+Fiona==1.9.3
+fonttools==4.43.1
+GDAL==3.5.1
+gitdb==4.0.11
+GitPython==3.1.37
+importlib-metadata==6.8.0
+jmespath==1.0.1
+joblib==1.3.2
+kiwisolver==1.4.5
+marshmallow==3.20.1
+marshmallow-enum==1.5.1
+matplotlib==3.5.1
+munch==4.0.0
+mypy-extensions==1.0.0
+numpy==1.24.3
+packaging==23.2
+pandas==2.0.3
+papipyplug==2023.10.3
+Pillow==10.1.0
+# pydsstools==2.2
+pyparsing==3.1.1
+pyproj==3.5.0
+python-dateutil==2.8.2
+python-dotenv==0.20.0
+pytz==2023.3.post1
+rasterio==1.3.9
+rioxarray==0.11.1
+s3transfer==0.5.2
+scikit-learn==1.2.2
+scipy==1.10.1
+shapely==2.0.1
+six==1.16.0
+smmap==5.0.1
+snuggs==1.4.7
+threadpoolctl==3.2.0
+typing-extensions==4.8.0
+typing-inspect==0.9.0
+tzdata==2023.3
+urllib3==1.26.18
+xarray==2023.1.0
+zipp==3.17.0
