+++ plugins/temp_precip/entrypoint.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+echo 'Temperature and precipitation DSS extraction plugin for AORC .zarr data'
