+++ plugins/temp_precip/requirements.txt
@@ -1,0 +1,51 @@
+affine==2.4.0
+aiobotocore==2.5.0
+aiohttp==3.8.4
+aioitertools==0.11.0
+aiosignal==1.3.1
+async-timeout==4.0.3
+attrs==23.1.0
+boto3==1.22.2
+botocore==1.29.76
+certifi==2023.7.22
+charset-normalizer==3.3.1
+click==8.1.7
+click-plugins==1.1.1
+cligj==0.7.2
+frozenlist==1.4.0
+fsspec==2023.5.0
+GDAL==3.5.1
+gitdb==4.0.11
+GitPython==3.1.37
+idna==3.4
+importlib-resources==6.1.0
+jmespath==1.0.1
+jsonschema==4.17.3
+multidict==6.0.4
+numpy==1.24.3
+packaging==23.2
+pandas==2.0.3
+papipyplug==2023.10.3
+pkgutil-resolve-name==1.3.10
+# pydsstools==2.2
+pyparsing==3.1.1
+pyproj==3.5.0
+pyrsistent==0.20.0
+python-dateutil==2.8.2
+python-dotenv==0.20.0
+pytz==2023.3.post1
+rasterio==1.3.9
+rioxarray==0.11.1
+s3fs==2023.5.0
+s3transfer==0.5.2
+shapely==2.0.1
+six==1.16.0
+smmap==5.0.1
+snuggs==1.4.7
+typing-extensions==4.8.0
+tzdata==2023.3
+urllib3==1.26.18
+wrapt==1.15.0
+xarray==2022.3.0
+yarl==1.9.2
+zipp==3.17.0
